% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_function.R
\docType{package}
\name{sfc-package}
\alias{sfc-package}
\title{sfc: A package for substance flow computation.}
\description{
The sfc package provides an important functions sfc() to compute
   the substance flow with the input files --- "data" and "model".
   If sample.size is set more than 1, uncertainty analysis will be
   executed while the distributions and parameters are supplied in
   the file "data".
}

